pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iam-dashboard"
version = "0.1.0"
description = "SaaS platform for autonomous legal agents"
requires-python = ">=3.12"
dependencies = [
    "nicegui[plotly]>=2.21.0",
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.41",
    "alembic>=1.13.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "google-genai>=0.1.0",
    "llama-index>=0.12.52",
    "llama-index-embeddings-gemini>=0.3.0",
    "llama-index-readers-file>=0.2.0",
    "llama-index-vector-stores-postgres>=0.2.0",
    "agno>=1.7.5",
    "pymupdf>=1.26.0",
    "pytesseract>=0.3.10",
    "opencv-python-headless>=4.8.0",
    "pillow>=10.0.0",
    "dependency-injector>=4.48.1",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.2",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "psutil>=5.9.0",
    "reportlab>=4.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[dependency-groups]
dev = [
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
    "types-pyyaml>=6.0.12.20250516",
    "types-requests>=2.32.4.20250611",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
